requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aiofiles==24.1.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
rlottie-python==1.0.1
Pillow==10.4.0
asyncio-throttle==1.0.2
tenacity==8.5.0
certifi==2024.7.4